Copy the distribution instead of svn exporting it
When we create a distribution, we used to use svn export to make sure
that the final product didn't have ".svn" folder junk in it. Now that
we are using git, we can't use "svn export" to avoid extra version
control specific files. Copying just the folders we care about to make
the distribution should be good enough.