Explain a bit more about dmaCopyHalfWords
Explain that dmaCopyHalfWords lets us specify the DMA channel to use.
This helps the description flow better, since it is otherwise
confusing that we'd talk about the DMA channel unless it was already
known that dmaCopyHalfWords required us to specify it; making such an
assumption in an "/Introduction/ to Nintendo DS Programming" is very
bad.