github: Don't publish from PRs
Only publish from the master branch. Do not upload artifacts or publish
them unless we are running the job as a result of the master branch
updating. This keeps building the manual for all PRs raised, so that the
PR submitter can see if their changes break the build or not, without
actually publishing the PR's changes to the production copy of the
manual.